feat: refactor library event and book entities, update DTOs, and enhance validation

+++ library-events-consumer-v2/src/main/java/com/learnkafka/dto/BookResponseDto.java
@@ -3,10 +3,10 @@ package com.learnkafka.dto;
 import java.time.LocalDateTime;
 
 public record BookResponseDto(
-        Long bookId,
+        Integer bookId,
         String bookName,
         String bookAuthor,
-        Long libraryEventId,
+        Integer libraryEventId,
         LocalDateTime createdAt,
         LocalDateTime updatedAt
 ) {
