+++ library-events-consumer-v2/src/main/java/com/learnkafka/repository/LibraryEventRepository.java
@@ -1,8 +1,8 @@
 package com.learnkafka.repository;
 
-import com.learnkafka.domain.LibraryEvent;
+import com.learnkafka.entity.LibraryEvent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LibraryEventRepository extends JpaRepository<LibraryEvent, Long> {
+public interface LibraryEventRepository extends JpaRepository<LibraryEvent, Integer> {
 }
